+++ Ruddarr/Views/ContentView.swift
@@ -44,6 +44,10 @@ struct ContentView: View {
         .tabViewCustomization($tabCustomization)
         .tabBarMinimizeBehavior(.never)
         .onAppear {
+            if !isRunningIn(.preview) {
+                dependencies.router.selectedTab = settings.tab
+            }
+
             UITabBarItem.appearance().badgeColor = UIColor(settings.theme.tint)
         }
         .onBecomeActive(perform: handleScenePhaseChange)