+++ Ruddarr/Views/Movies/MovieContextMenu.swift
@@ -3,16 +3,11 @@ import SwiftUI
 struct MovieContextMenu: View {
     var movie: Movie
 
-    @Environment(AppSettings.self) private var settings
     @Environment(RadarrInstance.self) private var instance
 
     var body: some View {
         Group {
             MovieLinks(movie: movie)
-
-            if movie.exists, let config = settings.instanceById(instance.id) {
-                InstanceWebLink(instance: config, path: movie.webPath)
-            }
 
             if movie.exists {
                 Divider()