+++ Ruddarr/Views/Movies/MovieView.swift
@@ -132,10 +132,6 @@ struct MovieView: View {
             }
 
             MovieLinks(movie: movie)
-
-            if let config = settings.instanceById(instance.id) {
-                InstanceWebLink(instance: config, path: movie.webPath)
-            }
         }
     }