+++ Ruddarr/Views/Series/SeasonView.swift
@@ -230,12 +230,6 @@ struct SeasonView: View {
                 }
 
                 Section {
-                    if let config = settings.instanceById(instance.id) {
-                        InstanceWebLink(instance: config, path: series.webPath)
-                    }
-                }
-
-                Section {
                     deleteSeasonButton
                 }
             } label: {