+++ Ruddarr/Views/Series/SeriesItemView.swift
@@ -89,10 +89,6 @@ struct SeriesDetailView: View {
 
                 Section {
                     SeriesLinks(series: series)
-
-                    if let config = settings.instanceById(instance.id) {
-                        InstanceWebLink(instance: config, path: series.webPath)
-                    }
                 }
 
                 Section {